+++ docs/conf.py
@@ -258,3 +258,4 @@ intersphinx_mapping['glue_jupyter'] = ('https://glue-jupyter.readthedocs.io/en/s
 intersphinx_mapping['regions'] = ('https://astropy-regions.readthedocs.io/en/stable/', None)
 intersphinx_mapping['skimage'] = ('https://scikit-image.org/docs/stable/', None)
 intersphinx_mapping['specutils'] = ('https://specutils.readthedocs.io/en/stable/', None)
+intersphinx_mapping['photutils'] = ('https://photutils.readthedocs.io/en/stable/', None)