+++ jdaviz/configs/imviz/tests/test_simple_aper_phot.py
@@ -3,7 +3,6 @@ from astropy import units as u
 from astropy.tests.helper import assert_quantity_allclose
 from numpy.testing import assert_allclose, assert_array_equal
 
-from jdaviz.configs.imviz.plugins.aper_phot_simple.aper_phot_simple import SimpleAperturePhotometry
 from jdaviz.configs.imviz.tests.utils import BaseImviz_WCS_WCS, BaseImviz_WCS_NoWCS
 
 
@@ -12,7 +11,7 @@ class TestSimpleAperPhot(BaseImviz_WCS_WCS):
         self.imviz.link_data(link_type='wcs')  # They are dithered by 1 pixel on X
         self.imviz._apply_interactive_region('bqplot:circle', (0, 0), (9, 9))  # Draw a circle
 
-        phot_plugin = SimpleAperturePhotometry(app=self.imviz.app)
+        phot_plugin = self.imviz.app.get_tray_item_from_name('imviz-aper-phot-simple')
 
         # Populate plugin menu items.
         phot_plugin._on_viewer_data_changed()
@@ -56,24 +55,36 @@ class TestSimpleAperPhot(BaseImviz_WCS_WCS):
         tbl = self.imviz.get_aperture_photometry_results()
         assert len(tbl) == 2
         assert tbl.colnames == [
-            'id', 'xcenter', 'ycenter', 'sky_center', 'background', 'npix', 'aperture_sum',
-            'pixarea_tot', 'aperture_sum_counts', 'aperture_sum_counts_err', 'counts_fac',
-            'aperture_sum_mag', 'flux_scaling', 'mean', 'stddev', 'median', 'min', 'max',
+            'id', 'xcentroid', 'ycentroid', 'sky_centroid', 'background', 'sum',
+            'sum_aper_area', 'pixarea_tot', 'aperture_sum_counts', 'aperture_sum_counts_err',
+            'counts_fac', 'aperture_sum_mag', 'flux_scaling', 'min', 'max', 'mean', 'median',
+            'mode', 'std', 'mad_std', 'var', 'biweight_location', 'biweight_midvariance',
+            'fwhm', 'semimajor_sigma', 'semiminor_sigma', 'orientation', 'eccentricity',
             'data_label', 'subset_label', 'timestamp']
         assert_array_equal(tbl['id'], [1, 2])
         assert_allclose(tbl['background'], 0)
-        assert_quantity_allclose(tbl['npix'], 63.617251235193315 * u.pix)
+        assert_quantity_allclose(tbl['sum_aper_area'], 63.617251 * (u.pix * u.pix))
         assert_array_equal(tbl['pixarea_tot'], None)
         assert_array_equal(tbl['aperture_sum_counts'], None)
         assert_array_equal(tbl['aperture_sum_counts_err'], None)
         assert_array_equal(tbl['counts_fac'], None)
         assert_array_equal(tbl['aperture_sum_mag'], None)
         assert_array_equal(tbl['flux_scaling'], None)
-        assert_allclose(tbl['mean'], 1)
-        assert_allclose(tbl['stddev'], 0)
-        assert_allclose(tbl['median'], 1)
         assert_allclose(tbl['min'], 1)
         assert_allclose(tbl['max'], 1)
+        assert_allclose(tbl['mean'], 1)
+        assert_allclose(tbl['median'], 1)
+        assert_allclose(tbl['mode'], 1)
+        assert_allclose(tbl['std'], 0)
+        assert_allclose(tbl['mad_std'], 0)
+        assert_allclose(tbl['var'], 0)
+        assert_allclose(tbl['biweight_location'], 1)
+        assert_allclose(tbl['biweight_midvariance'], 0)
+        assert_quantity_allclose(tbl['fwhm'], 5.15018758 * u.pix)
+        assert_quantity_allclose(tbl['semimajor_sigma'], 2.18708329 * u.pix)
+        assert_quantity_allclose(tbl['semiminor_sigma'], 2.18708329 * u.pix)
+        assert_quantity_allclose(tbl['orientation'], 0 * u.deg)
+        assert_quantity_allclose(tbl['eccentricity'], 0)
         assert_array_equal(tbl['data_label'], ['has_wcs_1[SCI,1]', 'has_wcs_2[SCI,1]'])
         assert_array_equal(tbl['subset_label'], ['Subset 1', 'Subset 1'])
         assert tbl['timestamp'].scale == 'utc'
@@ -81,12 +92,12 @@ class TestSimpleAperPhot(BaseImviz_WCS_WCS):
         # BUG: https://github.com/glue-viz/glue-astronomy/issues/52
         # Sky should have been the same and the pix different, but not until bug is fixed.
         # The aperture sum might be different too if mask is off limit in second image.
-        assert_quantity_allclose(tbl['xcenter'], 4.5 * u.pix)
-        assert_quantity_allclose(tbl['ycenter'], 4.5 * u.pix)
-        sky = tbl['sky_center']
+        assert_quantity_allclose(tbl['xcentroid'], 4.5 * u.pix)
+        assert_quantity_allclose(tbl['ycentroid'], 4.5 * u.pix)
+        sky = tbl['sky_centroid']
         assert_allclose(sky.ra.deg, [337.518943, 337.519241])
         assert_allclose(sky.dec.deg, [-20.832083, -20.832083])
-        assert_allclose(tbl['aperture_sum'], 63.61725123519332)
+        assert_allclose(tbl['sum'], 63.61725123519332)
 
         # Make sure it also works on an ellipse subset.
         self.imviz._apply_interactive_region('bqplot:ellipse', (0, 0), (9, 4))
@@ -98,13 +109,13 @@ class TestSimpleAperPhot(BaseImviz_WCS_WCS):
         tbl = self.imviz.get_aperture_photometry_results()
         assert len(tbl) == 3  # New result is appended
         assert tbl[-1]['id'] == 3
-        assert_quantity_allclose(tbl[-1]['xcenter'], 4.5 * u.pix)
-        assert_quantity_allclose(tbl[-1]['ycenter'], 2 * u.pix)
-        sky = tbl[-1]['sky_center']
+        assert_quantity_allclose(tbl[-1]['xcentroid'], 4.5 * u.pix)
+        assert_quantity_allclose(tbl[-1]['ycentroid'], 2 * u.pix)
+        sky = tbl[-1]['sky_centroid']
         assert_allclose(sky.ra.deg, 337.51894336144454)
         assert_allclose(sky.dec.deg, -20.832777499255897)
-        assert_quantity_allclose(tbl[-1]['npix'], 28.274334 * u.pix)
-        assert_allclose(tbl[-1]['aperture_sum'], 28.274334)
+        assert_quantity_allclose(tbl[-1]['sum_aper_area'], 28.274334 * (u.pix * u.pix))
+        assert_allclose(tbl[-1]['sum'], 28.274334)
         assert_allclose(tbl[-1]['mean'], 1)
         assert tbl[-1]['data_label'] == 'has_wcs_1[SCI,1]'
         assert tbl[-1]['subset_label'] == 'Subset 2'
@@ -121,13 +132,13 @@ class TestSimpleAperPhot(BaseImviz_WCS_WCS):
         tbl = self.imviz.get_aperture_photometry_results()
         assert len(tbl) == 4  # New result is appended
         assert tbl[-1]['id'] == 4
-        assert_quantity_allclose(tbl[-1]['xcenter'], 4.5 * u.pix)
-        assert_quantity_allclose(tbl[-1]['ycenter'], 4.5 * u.pix)
-        sky = tbl[-1]['sky_center']
-        assert_allclose(sky.ra.deg, 337.518943)
-        assert_allclose(sky.dec.deg, -20.832083)
-        assert_quantity_allclose(tbl[-1]['npix'], 81 * u.pix)
-        assert_allclose(tbl[-1]['aperture_sum'], 0)
+        assert np.isnan(tbl[-1]['xcentroid'])
+        assert np.isnan(tbl[-1]['ycentroid'])
+        sky = tbl[-1]['sky_centroid']
+        assert np.isnan(sky.ra.deg)
+        assert np.isnan(sky.dec.deg)
+        assert_quantity_allclose(tbl[-1]['sum_aper_area'], 81 * (u.pix * u.pix))
+        assert_allclose(tbl[-1]['sum'], 0)
         assert_allclose(tbl[-1]['mean'], 0)
         assert tbl[-1]['data_label'] == 'has_wcs_1[SCI,1]'
         assert tbl[-1]['subset_label'] == 'Subset 3'
@@ -147,7 +158,7 @@ class TestSimpleAperPhot_NoWCS(BaseImviz_WCS_NoWCS):
     def test_plugin_no_wcs(self):
         # Most things already tested above, so not re-tested here.
         self.imviz._apply_interactive_region('bqplot:circle', (0, 0), (9, 9))  # Draw a circle
-        phot_plugin = SimpleAperturePhotometry(app=self.imviz.app)
+        phot_plugin = self.imviz.app.get_tray_item_from_name('imviz-aper-phot-simple')
         phot_plugin._on_viewer_data_changed()
 
         phot_plugin.data_selected = 'has_wcs[SCI,1]'
@@ -160,4 +171,4 @@ class TestSimpleAperPhot_NoWCS(BaseImviz_WCS_NoWCS):
         phot_plugin.vue_do_aper_phot()
         tbl = self.imviz.get_aperture_photometry_results()
         assert len(tbl) == 1  # Old table discarded due to incompatible column
-        assert_array_equal(tbl['sky_center'], None)
+        assert_array_equal(tbl['sky_centroid'], None)